@@ -6,27 +6,26 @@
 
     <groupId>com.github.ixtf.japp</groupId>
     <artifactId>bom</artifactId>
-    <version>1.0.0</version>
+    <version>3.0.0</version>
     <packaging>pom</packaging>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>11</maven.compiler.source>
-        <maven.compiler.target>11</maven.compiler.target>
-        <jappVersion>1.0.0</jappVersion>
-        <vertxVersion>4.0.3</vertxVersion>
-        <micrometerVersion>1.6.5</micrometerVersion>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <jappVersion>3.0.0</jappVersion>
+        <vertxVersion>4.1.2</vertxVersion>
+        <micrometerVersion>1.7.3</micrometerVersion>
         <slf4jVersion>1.7.30</slf4jVersion>
         <logbackVersion>1.2.3</logbackVersion>
-        <jacksonVersion>2.12.3</jacksonVersion>
-        <reactorVersion>2020.0.6</reactorVersion>
+        <jacksonVersion>2.12.4</jacksonVersion>
+        <reactorVersion>2020.0.10</reactorVersion>
         <guiceVersion>5.0.1</guiceVersion>
         <jaegerVersion>1.6.0</jaegerVersion>
-        <keycloakVersion>11.0.3</keycloakVersion>
-        <droolsVersion>7.51.0.Final</droolsVersion>
+        <keycloakVersion>13.0.0</keycloakVersion>
+        <droolsVersion>7.56.0.Final</droolsVersion>
         <grpcVersion>1.28.1</grpcVersion>
         <protobufVersion>3.11.4</protobufVersion>
-        <netifiVersion>1.6.10</netifiVersion>
     </properties>
 
     <dependencyManagement>
@@ -54,11 +53,6 @@
             <dependency>
                 <groupId>com.github.ixtf.japp</groupId>
                 <artifactId>mongo</artifactId>
-                <version>${jappVersion}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.github.ixtf.japp</groupId>
-                <artifactId>persistence-mongo</artifactId>
                 <version>${jappVersion}</version>
             </dependency>
             <dependency>
@@ -126,12 +120,12 @@
             <dependency>
                 <groupId>com.graphql-java</groupId>
                 <artifactId>graphql-java</artifactId>
-                <version>16.2</version>
+                <version>17.1</version>
             </dependency>
             <dependency>
                 <groupId>com.graphql-java</groupId>
                 <artifactId>graphql-java-extended-scalars</artifactId>
-                <version>16.0.0</version>
+                <version>17.0</version>
             </dependency>
             <dependency>
                 <groupId>net.bytebuddy</groupId>
@@ -225,14 +219,6 @@
                 <version>0.3.0</version>
                 <!--                <type>pom</type>-->
                 <!--                <scope>import</scope>-->
-            </dependency>
-
-            <dependency>
-                <groupId>com.netifi</groupId>
-                <artifactId>netifi-bom</artifactId>
-                <version>${netifiVersion}</version>
-                <type>pom</type>
-                <scope>import</scope>
             </dependency>
         </dependencies>
     </dependencyManagement>
