@@ -2,15 +2,18 @@ package com.github.ixtf.api;
 
 import java.lang.annotation.*;
 
-@Target({ElementType.METHOD})
+import static com.github.ixtf.api.GraphqlAction.Type.QUERY;
+
+@Target({ElementType.METHOD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 public @interface GraphqlAction {
-    String type();
+    Type type() default QUERY;
 
     String action();
 
     enum Type {
-        QUERY
+        QUERY,
+        MUTATION,
     }
 }
