@@ -1,9 +1,9 @@
 package com.github.ixtf.codec;
 
-import java.io.UnsupportedEncodingException;
 import java.math.BigInteger;
+import java.nio.charset.StandardCharsets;
 
-public class Base58 {
+public final class Base58 {
 
     public static final char[] ALPHABET = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz".toCharArray();
     private static final int[] INDEXES = new int[128];
@@ -59,11 +59,7 @@ public class Base58 {
         }
 
         byte[] output = copyOfRange(temp, j, temp.length);
-        try {
-            return new String(output, "US-ASCII");
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException(e);  // Cannot happen.
-        }
+        return new String(output, StandardCharsets.US_ASCII);
     }
 
     public static byte[] decode(String input) throws IllegalArgumentException {
