@@ -3,14 +3,18 @@ package com.github.ixtf.mongo;
 import com.github.ixtf.J;
 import com.github.ixtf.data.EntityDTO;
 import com.github.ixtf.persistence.IEntity;
+import com.github.ixtf.persistence.Sort;
+import com.mongodb.client.model.Sorts;
 import com.mongodb.reactivestreams.client.FindPublisher;
 import com.mongodb.reactivestreams.client.MongoClient;
 import com.mongodb.reactivestreams.client.MongoCollection;
 import com.mongodb.reactivestreams.client.MongoDatabase;
 import org.bson.BsonDocument;
+import org.bson.BsonDocumentReader;
 import org.bson.BsonDocumentWriter;
 import org.bson.Document;
 import org.bson.codecs.Codec;
+import org.bson.codecs.DecoderContext;
 import org.bson.codecs.EncoderContext;
 import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.conversions.Bson;
@@ -19,26 +23,65 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.security.Principal;
+import java.util.Collection;
 import java.util.Optional;
-import java.util.function.Function;
 import java.util.stream.Stream;
 
 import static com.github.ixtf.guice.GuiceModule.getInstance;
 import static com.mongodb.client.model.Filters.and;
 import static com.mongodb.client.model.Filters.eq;
 import static java.util.Optional.ofNullable;
+import static java.util.stream.Collectors.toUnmodifiableList;
 
 public class Jmongo {
     public static final String ID_COL = "_id";
     public static final String DELETED_COL = "deleted";
     public static final Mono<Bson> DELETED_FILTER$ = Mono.fromCallable(() -> eq(DELETED_COL, false));
+    public static final int DEFAULT_BATCH_SIZE = 10_000;
+
+    public static Mono<Bson> $match(Flux<Bson> condition$) {
+        return Flux.merge(condition$, DELETED_FILTER$).collectList().map(it -> new Document("$match", and(it)));
+    }
+
+    public static Bson $skip(int skip) {
+        return new Document("$skip", skip);
+    }
+
+    public static Bson $limit(int limit) {
+        return new Document("$limit", limit);
+    }
+
+    public static Bson $count() {
+        return $count("count");
+    }
+
+    public static Bson $count(String field) {
+        return new Document("$count", field);
+    }
+
+    public static Optional<Bson> $sortOpt(Collection<Sort> sorts) {
+        return sortOpt(sorts).map(it -> new Document("$sort", it));
+    }
+
+    public static Optional<Bson> sortOpt(Collection<Sort> sorts) {
+        final var collect = ofNullable(sorts).filter(J::nonEmpty).stream().flatMap(Collection::stream).map(it -> {
+            final var id = it.getId();
+            switch (it.getStart()) {
+                case asc:
+                    return Sorts.ascending(id);
+                default:
+                    return Sorts.descending(id);
+            }
+        }).collect(toUnmodifiableList());
+        return Optional.of(collect).filter(J::nonEmpty).map(Sorts::orderBy);
+    }
 
     public MongoUnitOfWork uow() {
         return new MongoUnitOfWork(this);
     }
 
     public <T> Mono<T> find(MongoCollection<T> collection, Object id) {
-        return Flux.defer(() -> collection.find(eq(ID_COL, id))).next();
+        return Flux.defer(() -> collection.find(eq(ID_COL, id)).batchSize(1)).next();
     }
 
     public boolean exists(Class clazz, Object id) {
@@ -49,16 +92,16 @@ public class Jmongo {
         return exists(entity.getClass(), entity.getId());
     }
 
-    public <T> Mono<FindPublisher<T>> findPublisher(MongoCollection<T> collection, Publisher<Bson> filter$) {
+    public <T> FindPublisher<T> findPublisher(MongoCollection<T> collection, Publisher<Bson> filter$) {
         return Flux.defer(() -> filter$).collectList().map(filters -> {
             if (J.isEmpty(filters)) {
-                return collection.find();
+                return collection.find().batchSize(DEFAULT_BATCH_SIZE).allowDiskUse(true);
             } else if (filters.size() == 1) {
-                return collection.find(filters.get(0));
+                return collection.find(filters.get(0)).batchSize(DEFAULT_BATCH_SIZE).allowDiskUse(true);
             } else {
-                return collection.find(and(filters));
+                return collection.find(and(filters)).batchSize(DEFAULT_BATCH_SIZE).allowDiskUse(true);
             }
-        });
+        }).block();
     }
 
     public Mono<Long> countCollection(MongoCollection collection, Publisher<Bson> filter$) {
@@ -131,12 +174,19 @@ public class Jmongo {
         return bsonDocument;
     }
 
+    public <T> T fromBsonDocument(BsonDocument bsonDocument, Class<T> clazz) {
+        final var reader = new BsonDocumentReader(bsonDocument);
+        final var registry = getInstance(CodecRegistry.class);
+        final var codec = registry.get(clazz);
+        return codec.decode(reader, DecoderContext.builder().build());
+    }
+
     public <T extends MongoEntityBase> Flux<T> list(Class<T> clazz) {
-        return findPublisher(entityCollection(clazz), DELETED_FILTER$).flatMapMany(Function.identity());
+        return Flux.defer(() -> findPublisher(entityCollection(clazz), DELETED_FILTER$));
     }
 
     public <T> Flux<T> query(Class<T> clazz, Publisher<Bson> filter$) {
-        return findPublisher(entityCollection(clazz), Flux.merge(filter$, DELETED_FILTER$)).flatMapMany(Function.identity());
+        return Flux.defer(() -> findPublisher(entityCollection(clazz), Flux.merge(filter$, DELETED_FILTER$)));
     }
 
     public <T> Flux<T> query(Class<T> clazz, Stream<Bson> stream) {
@@ -147,32 +197,16 @@ public class Jmongo {
         return query(clazz, Mono.justOrEmpty(iterable).flatMapMany(Flux::fromIterable));
     }
 
-    public <T> Flux<T> query(Class<T> clazz, Bson... array) {
-        return query(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray));
-    }
-
-    public <T> Flux<T> query(Class<T> clazz, Optional<Bson>... array) {
-        return query(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray).flatMap(Mono::justOrEmpty));
-    }
-
     public <T> Flux<T> query(Class<T> clazz, Publisher<Bson> filter$, int skip, int limit) {
-        return findPublisher(entityCollection(clazz), Flux.merge(filter$, DELETED_FILTER$)).flatMapMany(it -> it.skip(skip).limit(limit).batchSize(limit));
+        return Flux.defer(() -> findPublisher(entityCollection(clazz), Flux.merge(filter$, DELETED_FILTER$)).skip(skip).limit(limit).batchSize(limit));
     }
 
     public <T> Flux<T> query(Class<T> clazz, Stream<Bson> stream, int skip, int limit) {
-        return query(clazz, Mono.justOrEmpty(stream).flatMapMany(Flux::fromStream), skip, limit);
+        return query(clazz, Flux.fromStream(stream), skip, limit);
     }
 
     public <T> Flux<T> query(Class<T> clazz, Iterable<Bson> iterable, int skip, int limit) {
-        return query(clazz, Mono.justOrEmpty(iterable).flatMapMany(Flux::fromIterable), skip, limit);
-    }
-
-    public <T> Flux<T> query(Class<T> clazz, int skip, int limit, Bson... array) {
-        return query(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray), skip, limit);
-    }
-
-    public <T> Flux<T> query(Class<T> clazz, int skip, int limit, Optional<Bson>... array) {
-        return query(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray).flatMap(Mono::justOrEmpty), skip, limit);
+        return query(clazz, Flux.fromIterable(iterable), skip, limit);
     }
 
     public Mono<Long> count(MongoCollection collection, Publisher<Bson> filter$) {
@@ -191,19 +225,9 @@ public class Jmongo {
         return count(clazz, Mono.justOrEmpty(iterable).flatMapMany(Flux::fromIterable));
     }
 
-    public Mono<Long> count(Class clazz, Bson... array) {
-        return count(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray));
-    }
-
-    public Mono<Long> count(Class clazz, Optional<Bson>... array) {
-        return count(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray).flatMap(Mono::justOrEmpty));
-    }
-
     public <T> Flux<T> autocomplete(Class<T> clazz, Publisher<Bson> filter$, int limit) {
-        return findPublisher(entityCollection(clazz), Flux.merge(filter$, DELETED_FILTER$)).flatMapMany(publisher -> {
-            final var max = Math.max(limit, 10);
-            return publisher.limit(max).batchSize(max);
-        });
+        final var max = Math.max(limit, 10);
+        return Flux.defer(() -> findPublisher(entityCollection(clazz), Flux.merge(filter$, DELETED_FILTER$)).limit(max).batchSize(max));
     }
 
     public <T> Flux<T> autocomplete(Class<T> clazz, Stream<Bson> stream, int limit) {
@@ -214,31 +238,16 @@ public class Jmongo {
         return autocomplete(clazz, Mono.justOrEmpty(iterable).flatMapMany(Flux::fromIterable), limit);
     }
 
-    public <T> Flux<T> autocomplete(Class<T> clazz, int limit, Bson... array) {
-        return autocomplete(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray), limit);
-    }
-
-    public <T> Flux<T> autocomplete(Class<T> clazz, int limit, Optional<Bson>... array) {
-        return autocomplete(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray).flatMap(Mono::justOrEmpty), limit);
-    }
-
     public <T> Flux<T> autocomplete(Class<T> clazz, Publisher<Bson> filter$) {
         return autocomplete(clazz, filter$, 10);
     }
 
     public <T> Flux<T> autocomplete(Class<T> clazz, Stream<Bson> stream) {
-        return autocomplete(clazz, Mono.justOrEmpty(stream).flatMapMany(Flux::fromStream));
+        return autocomplete(clazz, Flux.fromStream(stream));
     }
 
     public <T> Flux<T> autocomplete(Class<T> clazz, Iterable<Bson> iterable) {
-        return autocomplete(clazz, Mono.justOrEmpty(iterable).flatMapMany(Flux::fromIterable));
+        return autocomplete(clazz, Flux.fromIterable(iterable));
     }
 
-    public <T> Flux<T> autocomplete(Class<T> clazz, Bson... array) {
-        return autocomplete(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray));
-    }
-
-    public <T> Flux<T> autocomplete(Class<T> clazz, Optional<Bson>... array) {
-        return autocomplete(clazz, Mono.justOrEmpty(array).flatMapMany(Flux::fromArray).flatMap(Mono::justOrEmpty));
-    }
 }
